i fix

+++ kernel/src/c/mm/alloc.rs
@@ -1,10 +1,10 @@
 unsafe extern "C" {
-    pub fn malloc(size: ::core::ffi::c_uint) -> *mut ::core::ffi::c_void;
+    pub fn malloc(size: ::core::ffi::c_ulong) -> *mut ::core::ffi::c_void;
 }
 unsafe extern "C" {
     pub fn realloc(
         block: *mut ::core::ffi::c_void,
-        new: ::core::ffi::c_uint,
+        new: ::core::ffi::c_ulong,
     ) -> *mut ::core::ffi::c_void;
 }
 unsafe extern "C" {
